Two stage regression

Perform a two stage fitting procedure similar to the one proposed by Bates, S., & Tibshirani, R. (2019). Log‐ratio lasso: Scalable, sparse estimation for log‐ratio models. Biometrics, 75(2), 613-624. This function uses the selected components by trac or log-ratio and fits a sparse model based on all possible log-ratios.

Usage

second_stage(
  Z,
  A = NULL,
  y,
  additional_covariates = NULL,
  betas,
  topk = NULL,
  nfolds = 5,
  method = c("regr", "classif"),
  criterion = c("1se", "min"),
  alpha = 0
)

Arguments

Z

n by p matrix containing log(X) (see Z from trac)

A

p by (t_size-1) binary matrix giving tree structure (t_size is the total number of nodes and the -1 is because we do not include the root). Only needed for trac based models. (see A from trac). If A = NULL, a sparse-log contrast model is assumed.

y

n vector (response) (see y from trac)

additional_covariates

n by p' matrix containing additional covariates, (see additional_covariates from trac). Non-compositional components are currently not penalized by the lasso

betas

pre-screened coefficients (see output gamma from trac or output beta from sparse_log_contrast)

topk

maximum number of pre-screened coefficients to consider. Default NULL.

nfolds

number of folds

method

string which estimation method to use should be "regr" or "classif"

criterion

which criterion should be used to select the coefficients "1se" or "min"

alpha

nudge the model to select on a higher or lower level of the tree. Only relevant for trac based models.

Value

list with: log_ratios: betas for log ratios; index: dataframe with index of the pre selected coefficients and the log ratio name; A: taxonomic tree information; method: regression or classification; cv_glmnet: output of glmnet, useful for prediction; criterion: which criterion to be used to select lambda based on cv (cross validation)

Details

Fit trac or sparse-log contrast and extract the selected components.

Fit a second-stage based on the selected components.